requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
flow_matching==1.0.10
fonttools==4.56.0
fsspec==2025.3.0
hydra-core==1.3.2
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pymap2d==0.1.15
pyparsing==3.2.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0
torchdiffeq==0.2.5
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
wcwidth==0.2.13
wheel==0.45.1
torchdiffeq==0.2.5
einops==0.8.1
hydra-colorlog==1.2.0
seaborn==0.13.2
rich==14.0.0
POT==0.9.5
timm==1.0.15
diffusers==0.35.1
wandb==0.21.3
imageio==2.37.0